requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
pytest
pytest-django